@@ -10,7 +10,8 @@ repository: https://github.com/fokus-team/fokus
 version: 0.1.0-dev
 
 environment:
-    sdk: ">=2.6.0 <3.0.0"
+    sdk: ">=2.7.0 <3.0.0"
+    flutter: ">=1.17.0 <2.0.0"
 
 dependencies:
     flutter:
@@ -18,15 +19,25 @@ dependencies:
     flutter_localizations:
         sdk: flutter
 
+    fokus_auth:
+        path: fokus_auth
+
     mongo_dart:
         git: https://github.com/fokus-team/mongo_dart.git
 
-    firebase_core: ^0.4.0+9
-    firebase_analytics: ^5.0.11
+    firebase_core: ^0.5.0
+    firebase_analytics: ^6.0.0
     firebase_crashlytics: ^0.1.4+1
-    firebase_remote_config: ^0.3.0+3
-    firebase_auth: ^0.16.1
-    google_sign_in: ^4.5.1
+
+    firebase_messaging: ^6.0.16
+    onesignal_flutter:
+        git: https://github.com/fokus-team/OneSignal-Flutter-SDK.git
+    flutter_local_notifications:
+        git:
+            url: https://github.com/fokus-team/flutter_local_notifications.git
+            path: flutter_local_notifications
+
+    googleapis: ^0.55.0
 
     get_it: ^4.0.2
     bloc: ^6.0.1
@@ -42,7 +53,7 @@ dependencies:
     shared_preferences: ^0.5.8
     url_launcher: ^5.5.2
 
-    cupertino_icons: ^0.1.2
+    cupertino_icons: ^1.0.0
     lottie: ^0.4.0+1
     flutter_svg: ^0.18.0
     flutter_markdown: ^0.3.4
@@ -88,7 +99,6 @@ flutter:
         - assets/help/en/
         - assets/help/pl/
         - assets/image/avatar/
-        - assets/image/avatar/child/
         - assets/image/reward/
         - assets/image/badge/
         - assets/image/currency/
