@@ -4,7 +4,7 @@ import 'package:fokus/logic/auth/auth_bloc/authentication_bloc.dart';
 import 'package:fokus/logic/auth/child/child_auth_cubit_base.dart';
 import 'package:fokus/logic/auth/formz_state.dart';
 import 'package:fokus/model/db/user/user_role.dart';
-import 'package:fokus/utils/id_code_transformer.dart';
+import 'package:fokus_auth/fokus_auth.dart';
 import 'package:fokus/model/ui/auth/user_code.dart';
 
 part 'child_sign_in_state.dart';
