@@ -264,7 +264,7 @@ class _CaregiverCalendarPageState extends State<CaregiverCalendarPage> with Tick
 	        child: ItemCard(
 	          title: item.title,
 	          subtitle: AppLocales.of(context).translate(checked ? 'actions.selected' : 'actions.tapToSelect'),
-	          graphicType: GraphicAssetType.childAvatars,
+	          graphicType: AssetType.avatars,
 	          graphic: item.meta.avatar,
 	          graphicShowCheckmark: checked,
 	          graphicHeight: 44.0,
