@@ -1,12 +1,10 @@
-import 'dart:developer';
-
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:fokus/model/db/date/time_date.dart';
 import 'package:fokus/model/ui/gamification/ui_badge.dart';
 import 'package:fokus/services/app_locales.dart';
 import 'package:fokus/utils/app_paths.dart';
 import 'package:fokus/utils/dialog_utils.dart';
+import 'package:fokus/utils/icon_sets.dart';
 import 'package:fokus/utils/theme_config.dart';
 import 'package:fokus/widgets/app_navigation_bar.dart';
 import 'package:fokus/widgets/app_header.dart';
@@ -169,7 +167,7 @@ class _ChildAchievementsPageState extends State<ChildAchievementsPage> {
 										child: GestureDetector(
 											onTap: () => showBadgeDialog(context, badge),
 											child: SvgPicture.asset(
-												badgeIconSvgPath(badge.icon),
+												AssetType.badges.getPath(badge.icon),
 												width: (MediaQuery.of(context).size.width - 2 * AppBoxProperties.screenEdgePadding)/badgesPerShelf - 2 * badgeMargin
 											)
 										)
