@@ -293,7 +293,8 @@ class AppHeader extends StatelessWidget {
 	 Widget _userHeaderIcons(BuildContext context) {
 		return Row(
 			children: <Widget>[
-				headerIconButton(Icons.notifications, () => Navigator.of(context).pushNamed(AppPage.notificationsPage.name)),
+				// Supporting system notification only for now
+				// headerIconButton(Icons.notifications, () => Navigator.of(context).pushNamed(AppPage.notificationsPage.name)),
 				PopupMenuList(
 					lightTheme: true,
 					items: [
