@@ -3,6 +3,7 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:fokus/model/ui/gamification/ui_reward.dart';
 import 'package:fokus/services/app_locales.dart';
 import 'package:fokus/utils/app_paths.dart';
+import 'package:fokus/utils/icon_sets.dart';
 import 'package:fokus/utils/theme_config.dart';
 import 'package:fokus/widgets/buttons/rounded_button.dart';
 import 'package:fokus/widgets/chips/attribute_chip.dart';
@@ -61,7 +62,7 @@ class _RewardDialogState extends State<RewardDialog> with SingleTickerProviderSt
 									),
 									Padding(
 										padding: EdgeInsets.only(top: 10.0),
-										child: SvgPicture.asset(rewardIconSvgPath(widget.reward.icon), height: MediaQuery.of(context).size.width*0.3)
+										child: SvgPicture.asset(AssetType.rewards.getPath(widget.reward.icon), height: MediaQuery.of(context).size.width*0.3)
 									)
 								]
 							),
